demo_bidirection_streaming/requirements.txt
grpcio-tools
grpcio
protobuf
multithreading
python-json-config
pandas
numpy
math
